+++ WalletWasabi.Tests/UnitTests/MockIHttpClient.cs
@@ -23,17 +23,3 @@ public class MockIHttpClient : IHttpClient
 	public Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken = default) =>
 		OnSendAsync?.Invoke(request) ?? throw new NotImplementedException();
 }
-
-public static class MockHttpClientExtensions
-{
-	public static void SetupSequence(this MockHttpClient http, params Func<HttpResponseMessage>[] responses)
-	{
-		var callCounter = 0;
-		http.OnSendAsync = req =>
-		{
-			var responseFn = responses[callCounter];
-			Interlocked.Increment(ref callCounter);
-			return Task.FromResult(responseFn());
-		};
-	}
-}
